Move to separate folder AnimatedButtonTests
