+++ WalletWasabi.Tests/UnitTests/Fluent/Controls/AnimatedButtonTests/AnimatedButton001.axaml.cs
@@ -1,6 +1,6 @@
 ﻿using Avalonia.Controls;
 
-namespace WalletWasabi.Tests.UnitTests.Fluent.Controls;
+namespace WalletWasabi.Tests.UnitTests.Fluent.Controls.AnimatedButtonTests;
 
 public partial class AnimatedButton001 : Window
 {
