+++ WalletWasabi.Tests/UnitTests/Fluent/Controls/AnimatedButtonTests/AnimatedButtonTests.AnimatedButton_001.OSX.verified.png
